requirements.txt
Pillow==5.0.0
requests==2.17.3
